+++ docs/source/conf.py
@@ -404,6 +404,8 @@ nitpick_ignore = [
     ('py:class', 'FilterType'),
     ('py:class', 'XMLLike'),
     ('py:class', 'etree.XPathElementEvaluator'),
+    ('py:class', 'contextlib._GeneratorContextManager'),
+    ('py:data', 'masci_tools.io.parsers.fleur.fleur_outxml_parser.F')
 ]